% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_gene_cor}
\alias{calculate_gene_cor}
\title{Calculate Correlation Coefficient from an Expression Matrix}
\usage{
calculate_gene_cor(data, gene1, gene2, use.row = FALSE,
  method = "pearson")
}
\arguments{
\item{data}{a matrix or data.frame}

\item{gene1, gene2}{a character string gene name that's used as a column
name in the expression matrix}

\item{use.row}{by default FALSE, if TRUE, calculate_gene_cor will take
row names as gene names; if FALSE, column names will be seens as gene names}

\item{method}{a character string indicating which correlation coefficient to
compute (Options: "\code{pearson}" (default), "\code{kendall}",
or "\code{spearman}")}
}
\value{
a number
}
\description{
\code{calculate_gene_cor()} takes an expression matrix. Two genes will be
selected by column/row names and Pearson's correlation will be calculated
by \code{\link{cor}()}.
}
\examples{
genedf <- data.frame("a" = c(1:10),
                     "b" = c(6:15))
calculate_gene_cor(genedf, "a", "b")
}
