% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_conversion.R
\name{convert_to_genename}
\alias{convert_to_genename}
\title{Convert Gene Names or FlyBase Gene Numbers to Gene Names Specified in the
Database}
\usage{
convert_to_genename(x, db, normalize = TRUE, remove.dup = TRUE)
}
\arguments{
\item{x}{a character vector containing genes and FlyBase gene numbers}

\item{db}{a list from \code{make_database()}}

\item{normalize}{a logical expression. If set as TRUE, it will call
\code{\link{normalize_genename}()} to make consistent the converted gene
names.}
}
\description{
\code{convert_to_genename()} takes a character vector of gene names and
FlyBase gene numbers and uses the database list
\code{\link{make_database}()} generated to convert them to gene names that
are consistent with the version of the provided GTF file.
}
\examples{
# Make databases from dummy data
dummypath <- system.file("extdata", "dummy.gtf",
package = "genofeatutil")
testdb <- make_database(species = "test",
gtf.path = dummypath)

convert_to_genename("FBgn0086917", testdb)
}
