% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_conversion.R
\name{denormalize_genename}
\alias{denormalize_genename}
\title{Remove Prefix Added When Making Gene Names Legal Variable Names}
\usage{
denormalize_genename(gene)
}
\arguments{
\item{gene}{a character vector containing gene names}
}
\value{
a character vector containing converted gene names with prefix
removed
}
\description{
\code{denormalize_genename()} removes the prefix
\code{\link{normalize_genename}()} adds, and make the gene names more
readable for humans.
}
\examples{
gene <- normalize_genename("128up")
# [1] "gn_128up"
denormalize_genename(gene)
# [1] "128up"
}
