% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_conversion.R
\name{distinguish_fbgn}
\alias{distinguish_fbgn}
\title{Distinguish FlyBase Gene Numbers from Gene Names}
\usage{
distinguish_fbgn(x)
}
\arguments{
\item{x}{a character vector containing genes and FlyBase gene numbers}
}
\value{
a logical vector
}
\description{
Distinguish FlyBase Gene Numbers from Gene Names
}
