% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_conversion.R
\name{generate_fbid_version_table}
\alias{generate_fbid_version_table}
\title{Generate a Conversion Vector for FlyBase IDs from Older Versions}
\usage{
generate_fbid_version_table(db)
}
\arguments{
\item{db}{a list generated by \code{prepare_database()}}
}
\value{
a list without FBgn ID annotation table but with a derived
conversion vector
}
\description{
\code{generate_fbid_version_table()} takes the FBgn ID annotation table
from FlyBase, and prepare a conversion vector.
\code{generate_fbid_version_table()} takes a list generated by
\code{\link{prepare_database}()} and updates it by adding the conversion
vector and removing the raw FBgn ID annotation table.
}
