% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_conversion.R
\name{generate_flybase_sym}
\alias{generate_flybase_sym}
\title{Generate Conversion Tables for Symbols and Aliases to FlyBase ID}
\usage{
generate_flybase_sym(db)
}
\arguments{
\item{db}{a list generated by \code{\link{prepare_database}()}}
}
\value{
a list of 2 (a named vector that translate symbols to FB ids and
another that translates aliases to FB ids)
}
\description{
Generate Conversion Tables for Symbols and Aliases to FlyBase ID
}
