% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_conversion.R
\name{generate_gene_mapping}
\alias{generate_gene_mapping}
\title{Generate a Mapping Vector to Translate Gene IDs}
\usage{
generate_gene_mapping(db, ...)
}
\arguments{
\item{db}{a list generated by \code{\link{prepare_database}()}}

\item{...}{other arguments that are passed to \code{\link{get_gene_names}()}
(see below)}
}
\value{
a named character vector, in which the values are gene names and
the names are gene ids
}
\description{
\code{generate_gene_mapping()} takes gene names from a character vector, an
expression matrix / data frame, or a Seurat object, and generate a vector
that can be used to translate gene IDs to the gene names provided to
\code{generate_gene_mapping()}. The purpose of this translation is to make
sure different datasets are consistent in the gene names used during
analysis, and to thereby ensure comparsion between datasets or querying are
accurate.
}
