% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_flybase.R
\name{get_fbase_url}
\alias{get_fbase_url}
\title{Retrieving Download URLs for FlyBase ID conversion table and Symbol Synonym Table}
\usage{
get_fbase_url(version = NULL)
}
\arguments{
\item{version}{A character specifying the version desired (e.g., "FB2019_01".)}
}
\value{
A named list with 2 items ("fbid" and "syno")
}
\description{
\code{get_fbase_url()} calls RCurl to check FlyBase FTP release repository to
check available version. If not otherwise specified, it will generate dowdload
URLs for table for FlyBase ID and Synonym.
}
\examples{
# This retrieves the URLs for the up-to-date version
get_fbase_url()

# This gets the FB2019_01 version
get_fbase_url(version = "FB2019_01")
}
