% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_gene_names}
\alias{get_gene_names}
\alias{get_gene_names.default}
\alias{get_gene_names.Seurat}
\title{Get Gene Names from an Expression Matrix or a Seurat object}
\usage{
get_gene_names(x, ...)

\method{get_gene_names}{default}(x, use.row = FALSE, ...)

\method{get_gene_names}{Seurat}(x, assay = "RNA", ...)
}
\arguments{
\item{x}{an object (matrix, data frame, or Seuratv3 object) to extract gene
names from}

\item{...}{other arguments passing to S3 methods (see below)}

\item{use.row}{a logical constant indicating whether the gene names are in
row names or not}

\item{assay}{a character string indicating an assay in the Seurat object}
}
\value{
a character vector
}
\description{
\code{get_gene_names()} extracts gene names from a matrix, data frame, or
Seurat object. For a martix or data frame, you need to tell it whether the
gene names are in row names or column names. If it's a Seurat object,
you'll need to specify the assay you want.
}
\examples{
# Create a dummy expression matrix
expression_df <- data.frame(
"sample1" = c(1:10),
"sample2" = c(1:10),
"sample3" = c(1:10)
)

row.names(expression_df) <- c(paste("gene", c(1:10), sep = ""))

# Create Seurat object from the dummy data
seuratobj <- Seurat::CreateSeuratObject(counts = expression_df,
                                        assay = "RNA")
# Get gene names from a data frame
get_gene_names(expression_df, use.row = TRUE)

# Get gene names from a Seuratv3 object
get_gene_names(seuratobj, assay = "RNA")
}
