% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_from_SCENIC.R
\name{get_motif_info}
\alias{get_motif_info}
\title{Generate a Top Motif List Adjacent to Input Genes Based on SCENIC /
RcisTarget Database}
\usage{
get_motif_info(score_path, motif_path, number = NULL, threshold = NULL,
  genes.use = NULL, db)
}
\arguments{
\item{score_path}{a character string of the path to the scoring database
from SCENIC / RcisTarget}

\item{motif_path}{a character string of the path to the motif database
from SCENIC / RcisTarget}

\item{number}{a number of motifs with the highest score to leave from the
motif list}

\item{threshold}{a number of score threshold to filter the motifs}

\item{genes.use}{a character vector to leave only the genes of interest (
Usually the targets) in the output motif list}

\item{db}{a list from \code{make_database()}}
}
\description{
Generate a Top Motif List Adjacent to Input Genes Based on SCENIC /
RcisTarget Database
}
