% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{integrate_score}
\alias{integrate_score}
\title{Integrate Score Tables}
\usage{
integrate_score(..., column.name, na.zero = TRUE)
}
\arguments{
\item{...}{named arguments of dataframes; the names will be used as column
names in the result.}

\item{column.name}{a character string indicating the column name containing
the score to merge}

\item{na.zero}{a logical value; if TRUE, NAs will be replaced by 0}
}
\value{
a data frame containing integrated scores of interest
}
\description{
\code{integrate_score()} takes one specified column from multiple data
frames  containing the scores for prediction-target pair, and integrated
them into a summary table, in which each column represents one table, and
each row represents a prediction-target pair.
}
\examples{
# Generate dummy data
t1 <- data.frame("predictor" = c("tf_1", "tf_2", "tf_3"),
                 "target" = c("gene_1", "gene_2", "gene_3"),
                 "MSE" = c(1, 1, 0))
t2 <- data.frame("predictor" = c("tf_1", "tf_2", "tf_3"),
                 "target" = c("gene_1", "gene_2", "gene_3"),
                 "MSE" = c(1, 0, 1))
t3 <- data.frame("predictor" = c("tf_1", "tf_2", "tf_3", "tf_4"),
                 "target" = c("gene_1", "gene_2", "gene_3", "gene_1"),
                 "MSE" = c(0, 1, 1, 1))

integrated_table <- integrate_score(t1 = t1, t2 = t2, t3 = t3,
                                    column.name = "MSE")
}
