% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_conversion.R
\name{make_database}
\alias{make_database}
\title{Make an Reference for Gene Name Conversion}
\usage{
make_database(species = "dmel", gtf.path, version = NULL)
}
\arguments{
\item{species}{a character string describing the species of the genes.
(Options: "\code{dmel}")}

\item{gtf.path}{a character string containing the path to the GTF file
containing the input gene names (e.g., the GTF files used by
\code{cellranger count})}

\item{version}{a character specifying the version desired
(e.g., "FB2019_01".)}
}
\value{
a list containing multiple coversion tables and meta data
}
\description{
\code{make_database()} loads gene name / ID conversion tables from online
databases (e.g., FlyBase) and a user-provided GTF file that is used to
annotate the features of the NGS dataset of interest (e.g., the GTF file
STAR or \code{cellranger count} uses), and makes conversion tables from the
loaded data.
}
\examples{
dummypath <- system.file("extdata", "dummy.gtf",
package = "genofeatutil")
dmeldb <- make_database(species = "test",
                        gtf.path = dummypath)
}
