% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_from_SCENIC.R
\name{motif_to_tf}
\alias{motif_to_tf}
\title{Translate SCENIC Motif Names to Corresponding Transcription Factor Names}
\usage{
motif_to_tf(x, motif_db, db)
}
\arguments{
\item{x}{a list from \code{\link{get_motif_info}}}

\item{motif_db}{a character string containing the path the motif list from
SCENIC / RcisTarget}

\item{db}{a list from \code{make_database()}}
}
\value{
a list with the motifs names in each item of the list converted to
their corresponding transcription factor names
}
\description{
Translate SCENIC Motif Names to Corresponding Transcription Factor Names
}
