% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_conversion.R
\name{normalize_genename}
\alias{normalize_genename}
\title{Make Gene Names Legal Variable Names}
\usage{
normalize_genename(gene)
}
\arguments{
\item{gene}{a character vector containing gene names}
}
\value{
a character vector containing prefixed and converted gene names
}
\description{
It is common that gene names are saved as names of rows or columns of an
expression matrix. In R, this could create some problems because some gene
names are not legal variable names and thus will be converted when they are
saved as column names (which R assumes that they should be legal variable
names). This automatic confusion might interfere with name matching.
\code{normalize_genename()} converts gene names to make them legal variable
names, thereby make sure the gene names compared are consistently after
conversion. \code{normalize_genename()} will add a prefix (gn_) to gene
names to make gene names starting with a number legal, and then calls
\code{\link{make.names[base]}()} to convert illegal punctuations.
}
\seealso{
\code{\link{denormalize_genename}()}
}
