% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_conversion.R
\name{prepare_database}
\alias{prepare_database}
\title{Prepare an Reference for Gene Name Conversion}
\usage{
prepare_database(species = "dmel", gtf.path, version = NULL)
}
\arguments{
\item{species}{a character string describing the species of the genes.
(Options: "\code{dmel}")}

\item{gtf.path}{a character string containing the path to the GTF file
containing the input gene names (e.g., the GTF files used by
\code{cellranger count})}

\item{version}{}
}
\value{
a list containing multiple coversion tables and meta data
}
\description{
\code{prepare_database()} loads gene name / ID conversion tables from online
databases (e.g., FlyBase) and a user-provided GTF file that is used to
annotate the features of the NGS dataset of interest (e.g., the GTF file
STAR or \code{cellranger count} uses).
}
\examples{
dmeldb <- prepare_database(species = "dmel",
                           gtf.path = "extdata/dummy.gtf")
}
