% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_conversion.R
\name{update_fbgn}
\alias{update_fbgn}
\title{Update FlyBase Gene Numbers from an Older Version}
\usage{
update_fbgn(id, db)
}
\arguments{
\item{id}{a character vector containing FlyBase gene numbers}

\item{db}{a list from \code{make_database()}}
}
\value{
a character vector containing updated FlyBase gene numbers
}
\description{
\code{update_fbgn()} takes a character vector of FlyBase gene numbers and a
list from \code{\link{make_database}()}, and convert the vector according
to the list.
}
\examples{
dummypath <- system.file("extdata", "dummy.gtf",
package = "genofeatutil")
dmeldb <- make_database(species = "dmel",
                           gtf.path = dummypath)
update_fbgn("FBgn0032045", db = dmeldb)
}
