% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_process.R
\name{versaread}
\alias{versaread}
\title{Load RDS, RData, or CSV files}
\usage{
versaread(file, type)
}
\arguments{
\item{file}{The path of the desired file}

\item{type}{The type of file being loaded (Options: rds, rdata, or csv)}
}
\value{
a character vector or an R object (if it's an RData or RDS file)
}
\description{
\code{versaread()} is mainly for loading saved lists of gene names.
For .RData files containing one object or an .RDS file, \code{versaread()} loads and returns
the object for you to assign. For CSV files, it expects something containing 1 column
and with a header, and will ignore everything after the first column.
}
\examples{
saveRDS(iris, "~/iris.rds")
iris_reload <- versaread("~/iris.rds", type = "rds")
write.csv(iris$Species, "iris.csv")
iris_species <- versaread("iris.csv", type = "csv")
}
