% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_conversion.R
\name{generate_flybase_sym}
\alias{generate_flybase_sym}
\title{Title}
\usage{
generate_flybase_sym(db)
}
\arguments{
\item{db}{}
}
\value{
a list of 2 (a named vector that translate symbols to FB ids and
another that translates aliases to FB ids)
}
\description{
Title
}
