% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_mtry_series}
\alias{make_mtry_series}
\title{Generate a Tuning Series of \code{mtry} for \code{randomForest()}}
\usage{
make_mtry_series(x, y)
}
\arguments{
\item{x}{a predictor matrix}

\item{y}{a response matrix}
}
\value{
a numeric factor
}
\description{
\code{\link[randomForest]{randomForest}()} implements a default value for
mtry determined by the number of predictors used. To examine the influence
of mtry on prediction accuracy, \code{make_mtry_series()} takes predictor
and response matrices and generate a series of mtry for use in a tune grid.
Let default value be N, \code{make_mtry_series()} makes a numeric vector
consisting 10, N/2, N, 2N, and the number of predictors provided.
the number of predictors.
}
\examples{
make_mtry_series(x = iris[ , 1:4], y = iris[ , 5])
}
