% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_process.R
\name{versaread}
\alias{versaread}
\title{Load RDS, RData, or CSV Files}
\usage{
versaread(file, type)
}
\arguments{
\item{file}{a character string indicating the path of the desired file}

\item{type}{a character string indicating the type of file being loaded
(Options: \code{rds}, \code{rdata}, or \code{csv})}
}
\value{
a character vector or an R object (if it's an RData or RDS file)
}
\description{
\code{versaread()} is mainly for loading saved lists of gene names.
For .RData files containing one object or an .RDS file, \code{versaread()}
loads and returns the object for you to assign. For CSV files, it expects
something containing 1 column and with a header, and will ignore everything
after the first column.
}
\examples{
temprds <- tempfile(fileext = ".rds")
saveRDS(iris, file = temprds)
iris_reload <- versaread(temprds, type = "rds")
tempcsv<- tempfile(fileext = ".csv")
write.csv(iris$Species, file = tempcsv)
iris_species <- versaread(tempcsv, type = "csv")
}
